+++ lib/praxis/multipart/part.rb
@@ -11,6 +11,22 @@ module Praxis
       @filename = filename
     end
 
+    # Determine the content type of this response.
+    #
+    # @return [MediaTypeIdentifier]
+    def content_type
+      MediaTypeIdentifier.load(headers['Content-Type']).freeze
+    end
+
+    # Set the content type for this response.
+    # @todo DRY this out (also used in Response)
+    #
+    # @return [String]
+    # @param [String,MediaTypeIdentifier] identifier
+    def content_type=(identifier)
+      headers['Content-Type'] = MediaTypeIdentifier.load(identifier).to_s
+    end
+
     def status
       @headers['Status'].to_i
     end