+++ lib/praxis/types/media_type_common.rb
@@ -1,6 +1,7 @@
 module Praxis
   module Types
 
+    # Traits that are shared by MediaType and SimpleMediaType.
     module MediaTypeCommon
       extend ::ActiveSupport::Concern
 
@@ -18,10 +19,14 @@ module Praxis
           @description
         end
 
+        # Get or set the identifier of this media type.
+        #
+        # @deprecated this method is not deprecated, but its return type will change to MediaTypeIdentifier in Praxis 1.0
+        #
+        # @return [String] the string-representation of this type's identifier
         def identifier(identifier=nil)
-          return @identifier unless identifier
-          # TODO: parse the string and extract things like collection , and format type?...
-          @identifier = identifier
+          return @identifier.to_s unless identifier
+          (@identifier = MediaTypeIdentifier.load(identifier)).to_s
         end
       end