+++ spec/praxis/types/collection_spec.rb
@@ -14,7 +14,7 @@ describe Praxis::Collection do
   context '.of' do
     let(:media_type) do
       Class.new(Praxis::MediaType) do
-        identifier 'an-awesome-type'
+        identifier 'application/an-awesome-type'
       end
     end
 
@@ -22,7 +22,7 @@ describe Praxis::Collection do
       Praxis::Collection.of(media_type)
     end
 
-    its(:identifier) { should eq 'an-awesome-type;type=collection' }
+    its(:identifier) { should eq 'application/an-awesome-type; type=collection' }
 
     it 'sets the collection on the media type' do
       expect(media_type::Collection).to be(collection)